Add files via upload
Both the .crx and the original folder are included in this .zip file. However I have corrected the background image name from "ntp_background.png" to "ntp_background.PNG" which is written in the .json file. This has not caused issues on other operating systems however in the case of Linux where every file is case sensitive this causes the theme to not load from the Chrome webstore. The theme should now install on all operating systems without issue.